requirements.txt
groq
python-dotenv
openai
pymupdf
Pillow
pytesseract
flask